requirements.txt
click
aioredis
colorlog
booby-ng
msgpack-python